enhance cart page
