+++ src/types/product_Type.ts
@@ -27,14 +27,3 @@ export interface colorsVariants {
   images: string[];
   sizes?: string[];
 }
-
-export type cartProduct = {
-  id: number;
-  mainImage: string;
-
-  title: string;
-  price: number;
-  quantity: number;
-  color: string | undefined;
-  key: string;
-};
